Added default playlist logic to db layer. Working in playlist ui component.

+++ components/alexandria/src/main/java/com/frostwire/alexandria/db/PlaylistDB.java
@@ -47,15 +47,26 @@ public class PlaylistDB extends ObjectDB<Playlist> {
     }
 
     public void delete(Playlist obj) {
+        if (obj.getName().equals(LibraryDatabase.DEFAULT_PLAYLIST_NAME)) {
+            return;
+        }
         db.update("DELETE FROM PlaylistsPlaylistItems WHERE playlistId = " + obj.getId());
         db.update("DELETE FROM Playlists WHERE playlistId = " + obj.getId());
     }
 
     public List<PlaylistItem> getLibraryItems(Playlist playlist) {
-        List<List<Object>> result = db
-                .query("SELECT playlistItemId, filePath, fileName, fileSize, fileExtension, trackTitle, duration, artistName, albumName, coverArtPath "
-                        + "FROM PlaylistItems INNER JOIN PlaylistsPlaylistItems ON PlaylistItems.playlistItemId = PlaylistsPlaylistItems.playlistItemId "
-                        + "WHERE playlistId = " + playlist.getId());
+        String query;
+
+        if (playlist.getName().equals(LibraryDatabase.DEFAULT_PLAYLIST_NAME)) {
+            query = "SELECT playlistItemId, filePath, fileName, fileSize, fileExtension, trackTitle, duration, artistName, albumName, coverArtPath "
+                    + "FROM PlaylistItems";
+        } else {
+            query = "SELECT playlistItemId, filePath, fileName, fileSize, fileExtension, trackTitle, duration, artistName, albumName, coverArtPath "
+                    + "FROM PlaylistItems INNER JOIN PlaylistsPlaylistItems ON PlaylistItems.playlistItemId = PlaylistsPlaylistItems.playlistItemId "
+                    + "WHERE playlistId = " + playlist.getId();
+        }
+
+        List<List<Object>> result = db.query(query);
 
         List<PlaylistItem> items = new ArrayList<PlaylistItem>(result.size());
